% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_hello_roxygen.R
\name{my_hello_roxygen}
\alias{my_hello_roxygen}
\title{My Hello World Function using Roxygen}
\usage{
my_hello_roxygen(x)
}
\arguments{
\item{x}{The name of the person to say hi to}
}
\value{
The output from \code{\link{print}}
}
\description{
My Hello World Function using Roxygen
}
\examples{
hello("John")
\dontrun{
hello("Steve")
}
}
